# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "自控力API",
        "title": "Self_Control API",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "TAODEI",
            "email": "864978550@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "39.99.53.8:2333",
    "basePath": "/api/v1",
    "paths": {
...
        "/user": {
            "get": {
                "description": "获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "203": {
                        "description": "{\"error_code\":\"20001\", \"message\":\"Fail.\"}",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "接收新的User结构体来修改用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "需要修改的用户信息",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"修改成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "{\"error_code\":\"20001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "学号密码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录的用户信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"将student_id作为token保留\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "{\"error_code\":\"20001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Password or account wrong.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/user/goldhistory": {
            "get": {
                "description": "获取金币历史",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "金币历史",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/user/privacy/{id}": {
            "get": {
                "description": "判断该用户是否选择公开自己的打卡标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "隐私",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"默认为1 若要修改隐私 直接使用修改用户信息\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "203": {
                        "description": "未找到该用户"
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/user/punch": {
            "get": {
                "description": "获取我的打卡数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "我的打卡数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "error.Error": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "current_backdrop": {
                    "type": "integer"
                },
                "gold": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "privacy": {
                    "type": "integer"
                },
                "student_id": {
                    "type": "string"
                },
                "user_picture": {
                    "type": "string"
                }
            }
        }
    }
}